fix: DidRedClick and DidYellowClick
Replaced MainScreen.DidRedClick() and MainScreen.DidYellowClick() with a modified version of Flight's / old SRLs.

The one we were using had issues on the borders of the client and the current ones should always work. I made some modifications which in my opinion simplify them but the behavior shouldn't change from the ones Flight was using