fix(bankpin): read notes
- BankPin doesn't rely on bank being open to be considered successful anymore.
There's lot's of situations that is not opening the bank that require the bank pin. This were a bit of a pain in the ass to deal with because you would need to override BankPin.Enter() for each use case.
This is now solved and should work no matter what you are doing that asks for a bank pin.

- BankPin failure is handled much better now