fix: fixed a nasty bug introduced by the new unified iteminterface.
The bug in question was in TRSItemInterface.FindAll() and TRSItemInterface.ContainsAll().

This methods called `Unique()` which does not exist but can be called anyway for whatever reason. Probably exists in Simba1400 internally but is not exported to be available, idk but it was not working properly anyway and caused issues.