feat(bank): read notes
feat: bank interface now has a handler to the stupid jagex launcher ad when you open the bank sometimes. This is not very well tested because it's hard to stress test it since it won't show up in a while after being handled once but should work fine.
TLDR: Simba ad blocker.

refactor: ITEM_FINDER_IMAGES_VERSION was getting annoying to update each time I the images.zip was changed and it was a bit stupid to do it manually each time.
It now uses a hash of the zip file as a "version" folder which is much smarter although less human friendly I guess. To know the folder you can either delete all hash folders and run your script once or print `ItemFinder.IMAGES_PATH`.