fix(mainscreen): read notes
Red hit splats seem to have changed color for some reason. While fixing this I decided to change the method to detect hit splats to handle every scenario I knew of and should now detect:
- blue hit splats
- red hit splats
- max hit splats
- blue hit splat in GWD (they are different)
- red hit splats in GWD (they are different)
- max hit splats in GWD (they are different)

The colors used to do the CTS2Color are commented in the method in case more hit splats need to be added in the future.

Also solved a minor compile issue when trying to compile from within SRL-T files