feat: magic tab now uses the new spellfinder
While compatibility with older scripts was attempted, this could break things that were not expected.
Old methods that are not meant to be used anymore remain but were marked as deprecated.
With that said, magic should now be able to work with or without magic tab filters and with or without the spells resized. It also has a way to handle the "spell info" black box that pops up on hovering spells