fix(RemoteInput): read notes
Updated the plugin to included the latest version which finally has the ability to release the client.
SRL Adds on terminate instructions for RI to release the client early on, this should make people have way less issue when updating SRL.
People can still bypass this by force terminating a script (double click stop button) but since this is done early on in the SRL setup process they actually have to be fast about it. A second or so should be enough to release the client.
I've also added wasplib methods to enable/disable input, and before releasing the client we ensure:
- Input is enabled
- Client image is cleared of any drawing
- Client is released

BREAKING CHANGE: moving WaspLib remote input methods to SRL is a breaking change, WaspLib will also require an update