fix(Bank): TRSBank.FindItemBoundaries()
`TRSBank.FindItemBoundaries()` was finding extra boxes that would then be searched needlessly when trying to find items. The fix was done by ClubXj with a small performance tweak by me since it had 2 loops and only needed 1.
The difference is quite big the old method could easily find over 80% extra boxes than what was needed and they would be used to search for items, the new method should find the exact number of boxes with no extras. This was especially terrible in resizable.