fix(TRSWalker): TRSWalker.IsWalkable() was rolled back
Honestly, after reviewing my changes and the original I'm not even sure what I was trying to do and it only caused issues. The original looks good and should work fine.
I did however, increase the expansion of the color checking box a little bit because I found it used to click on black too often while running and click on the furthest tile possible