fix(TRSMapLoader): WebGraph generator optimizations
This still needs to be optimized further somehow, it's still not fast enough but here is a comparison of old vs new:

Code:
```
Map.SetupChunkEx([45, 53, 47, 50], [0]);
```

Results:
```
Old: Took 940 seconds to generate webgraph.
New: Took 852 seconds to generate webgraph.
```
The main reason this is slow and becomes slower the larger the map it's because of it's brute forcing nature.
Some ideas for the future is to try and not have so many "similar" paths... maybe based on paths angles.