feat(TRSMap): give control of the sample radius to people
This is more or less a power tool and unless you know TRSMap/TRSWalkerV1 you might not understand what this is for but this allows you to control the size of minimap sample taken for positioning.

This is very useful when you have issues where position gets thrown off a lot by a neighbouring chunk not loading when you are on a certain tile. By decreasing the sample radius you can make it so your minimap sample doesn't include the not yet loaded chunk.