feat:(TRSInterface): full rework
This is a squashed merge with a lot of commits, for every single change check each commit separately.
`TRSInterface` which was at the root of every interface in SRL and Wasplib was reworked to use variables instead of functions.
This is much more efficient and for some functions that used this interface functions as if they were variables the performance improvements can easily be above 200% faster.
A lot of things that were calculated on the fly everytime they were used are now only calculated once and cached for next uses.
If you didn't use interfaces bounds related functions before your script should just work.
BREAKING CHANGES: This is a breaking change and if you were using TRSInterface bounds related functions your script may break.